+++ test/OpenApiTests/ExistingOpenApiIntegration/FlightsController.cs
@@ -3,7 +3,7 @@ using JsonApiDotNetCore.Controllers;
 using JsonApiDotNetCore.Services;
 using Microsoft.Extensions.Logging;
 
-namespace OpenApiTests
+namespace OpenApiTests.ExistingOpenApiIntegration
 {
     public sealed class FlightsController : JsonApiController<Flight>
     {
